+++ validated/src/main/java/uk/ac/standrews/cs/digitising_scotland/verisim/populationStatistics/dataDistributionTables/statsKeys/SeparationStatsKey.java
@@ -24,7 +24,7 @@ import uk.ac.standrews.cs.digitising_scotland.verisim.dateModel.timeSteps.Compou
  */
 public class SeparationStatsKey extends StatsKey {
 
-    public SeparationStatsKey(Integer numberOfChildren, int forNPeople, CompoundTimeUnit consideredTimePeriod, Date currentDate) {
+    public SeparationStatsKey(Integer numberOfChildren, double forNPeople, CompoundTimeUnit consideredTimePeriod, Date currentDate) {
         super(numberOfChildren, forNPeople, consideredTimePeriod, currentDate);
     }
